Move files out of nested KachingPlugIn folder.
