+++ src/KachingPlugIn/EventHandlers/CatalogEventHandler.cs
