+++ src/KachingPlugIn/Factories/L10nStringFactory.cs
