+++ src/KachingPlugIn/Helpers/StringExtensions.cs
