+++ src/KachingPlugIn/KachingPlugIn.csproj
@@ -36,36 +36,36 @@
     <Compile Include="Configuration\KachingConfiguration.cs" />
     <Compile Include="Configuration\SystemMappingElement.cs" />
     <Compile Include="EnumerableExtensions.cs" />
-    <Compile Include="KachingPlugIn\CommerceMediaComparer.cs" />
-    <Compile Include="KachingPlugIn\EventHandlers\AssociationsEventHandler.cs" />
-    <Compile Include="KachingPlugIn\EventHandlers\CatalogEventHandler.cs" />
-    <Compile Include="KachingPlugIn\Factories\L10nStringFactory.cs" />
-    <Compile Include="KachingPlugIn\Factories\ProductFactory.cs" />
-    <Compile Include="KachingPlugIn\Helpers\APIFacade.cs" />
-    <Compile Include="KachingPlugIn\Helpers\StringExtensions.cs" />
-    <Compile Include="KachingPlugIn\Initialization\Initialization.cs" />
-    <Compile Include="KachingPlugIn\Models\AttributeTextValue.cs" />
-    <Compile Include="KachingPlugIn\Models\Dimension.cs" />
-    <Compile Include="KachingPlugIn\Models\DimensionValue.cs" />
-    <Compile Include="KachingPlugIn\Models\FieldMappingConfiguration.cs" />
-    <Compile Include="KachingPlugIn\Models\Folder.cs" />
-    <Compile Include="KachingPlugIn\Models\L10nString.cs" />
-    <Compile Include="KachingPlugIn\Models\L10nStringConverter.cs" />
-    <Compile Include="KachingPlugIn\Models\MarketPrice.cs" />
-    <Compile Include="KachingPlugIn\Models\MarketPriceConverter.cs" />
-    <Compile Include="KachingPlugIn\Models\PriceUnit.cs" />
-    <Compile Include="KachingPlugIn\Models\Product.cs" />
-    <Compile Include="KachingPlugIn\Models\ProductAsset.cs" />
-    <Compile Include="KachingPlugIn\Models\ProductMetadata.cs" />
-    <Compile Include="KachingPlugIn\Models\RecommendationGroup.cs" />
-    <Compile Include="KachingPlugIn\Models\Tag.cs" />
-    <Compile Include="KachingPlugIn\Models\UnitPricing.cs" />
-    <Compile Include="KachingPlugIn\Models\Variant.cs" />
-    <Compile Include="KachingPlugIn\Services\CategoryExportService.cs" />
-    <Compile Include="KachingPlugIn\Services\CategoryExportSingleton.cs" />
-    <Compile Include="KachingPlugIn\Services\IExportState.cs" />
-    <Compile Include="KachingPlugIn\Services\ProductExportService.cs" />
-    <Compile Include="KachingPlugIn\Services\ProductExportSingleton.cs" />
+    <Compile Include="CommerceMediaComparer.cs" />
+    <Compile Include="EventHandlers\AssociationsEventHandler.cs" />
+    <Compile Include="EventHandlers\CatalogEventHandler.cs" />
+    <Compile Include="Factories\L10nStringFactory.cs" />
+    <Compile Include="Factories\ProductFactory.cs" />
+    <Compile Include="Helpers\APIFacade.cs" />
+    <Compile Include="Helpers\StringExtensions.cs" />
+    <Compile Include="Initialization\Initialization.cs" />
+    <Compile Include="Models\AttributeTextValue.cs" />
+    <Compile Include="Models\Dimension.cs" />
+    <Compile Include="Models\DimensionValue.cs" />
+    <Compile Include="Models\FieldMappingConfiguration.cs" />
+    <Compile Include="Models\Folder.cs" />
+    <Compile Include="Models\L10nString.cs" />
+    <Compile Include="Models\L10nStringConverter.cs" />
+    <Compile Include="Models\MarketPrice.cs" />
+    <Compile Include="Models\MarketPriceConverter.cs" />
+    <Compile Include="Models\PriceUnit.cs" />
+    <Compile Include="Models\Product.cs" />
+    <Compile Include="Models\ProductAsset.cs" />
+    <Compile Include="Models\ProductMetadata.cs" />
+    <Compile Include="Models\RecommendationGroup.cs" />
+    <Compile Include="Models\Tag.cs" />
+    <Compile Include="Models\UnitPricing.cs" />
+    <Compile Include="Models\Variant.cs" />
+    <Compile Include="Services\CategoryExportService.cs" />
+    <Compile Include="Services\CategoryExportSingleton.cs" />
+    <Compile Include="Services\IExportState.cs" />
+    <Compile Include="Services\ProductExportService.cs" />
+    <Compile Include="Services\ProductExportSingleton.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
