+++ src/KachingPlugIn/Models/AttributeTextValue.cs
