+++ src/KachingPlugIn/Models/Dimension.cs
