+++ src/KachingPlugIn/Models/DimensionValue.cs
