+++ src/KachingPlugIn/Models/FieldMappingConfiguration.cs
