+++ src/KachingPlugIn/Models/Folder.cs
