+++ src/KachingPlugIn/Models/L10nString.cs
