+++ src/KachingPlugIn/Models/MarketPrice.cs
