+++ src/KachingPlugIn/Models/MarketPriceConverter.cs
