+++ src/KachingPlugIn/Models/ProductMetadata.cs
