+++ src/KachingPlugIn/Models/Variant.cs
