+++ src/KachingPlugIn/Services/CategoryExportSingleton.cs
