+++ src/KachingPlugIn/Services/IExportState.cs
