+++ src/KachingPlugIn/Services/ProductExportSingleton.cs
